+++ .gitignore
@@ -2,3 +2,6 @@ week5/lesson1_Flask/homework/DmytroMelnyk/TASK1/env_w5_t1/
 .idea/
 week3/OOP in Python/homework/DmytroMelnyk/utils/__pycache__/
 week4/lesson2_Request-Response_model/homework/DmytroMelnyk/__pycache__/
+week5/lesson1_Flask/homework/DmytroMelnyk/TASK1/envW5T1/
+week5/lesson1_Flask/homework/DmytroMelnyk/TASK2/task_2/__pycache__/
+
