--- week5/lesson1_Flask/homework/DmytroMelnyk/TASK1/.gitignor
@@ -1,22 +1,0 @@
-.idea/
-.idea/*
-
-lib/
-lib/*
-
-*.py[cod]
-*.pyc
-
-bin/
-bin/*
-
-share/
-share/*
-
-local/
-local/*
-
-include/
-include/*
-
-env/
