--- week5/lesson1_Flask/homework/DmytroMelnyk/TASK1/api/__init__.py
@@ -1,16 +1,0 @@
-from flask import Blueprint, render_template
-from flask.views import MethodView
-
-
-####################################################################
-
-class ApiView(MethodView):
-    def get(self):
-        return render_template("home.html")
-
-
-#####################################################################
-
-api = Blueprint('api', __name__, static_folder='./static', template_folder='./template')
-
-api.add_url_rule('/api', view_func=ApiView.as_view('api'))
