+++ week5/lesson1_Flask/homework/DmytroMelnyk/TASK1/app/MANIFEST.in
