+++ week5/lesson1_Flask/homework/DmytroMelnyk/TASK1/app/apicars/blueprint.py
@@ -1,0 +1,8 @@
+from flask import Blueprint, render_template
+
+apicars = Blueprint('apicars', __name__, template_folder='templates')
+
+
+@apicars.route('/')
+def index():
+    return render_template('apicars/home.html')
