+++ week5/lesson1_Flask/homework/DmytroMelnyk/TASK1/app/apicars/templates/apicars/home.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+        <!-- Required meta tags -->
+    <meta charset="utf-8">
+    {% if title %}
+        <title>Cars - {{title}}</title>
+    {% else %}
+        <title>Blueprint Cars</title>
+    {% endif %}
+</head>
+<body>
+    <h1>Blueprint Cars Home Page</h1>
+    <a href="/cars">GET ALL CARS</a>
+</body>
+</html>
