+++ week5/lesson1_Flask/homework/DmytroMelnyk/TASK1/app/app.py
@@ -1,0 +1,9 @@
+from flask import Flask
+from config import Configuration
+
+from apicars.blueprint import apicars
+
+app = Flask(__name__)
+app.config.from_object(Configuration)
+
+app.register_blueprint(apicars, url_prefix='/api')
