+++ week5/lesson1_Flask/homework/DmytroMelnyk/TASK1/app/main.py
@@ -1,0 +1,5 @@
+from app import app
+import routes
+
+if __name__ == "__main__":
+    app.run()
