+++ week5/lesson1_Flask/homework/DmytroMelnyk/TASK1/app/ourdb/ourdb.py
@@ -1,0 +1,7 @@
+carbase = [{"Model": "Chevelle Malibu", "Mark": "Chevrolet", "Horsepower": 130, "Year": "1970-01-01", "Origin": "USA"},
+           {"Model": "Skylark 320", "Mark": "Buick", "Horsepower": 165, "Year": "1970-01-01", "Origin": "USA"},
+           {"Model": "Satellite", "Mark": "Plymouth", "Horsepower": 150, "Year": "1970-01-01", "Origin": "USA"},
+           {"Model": "Torino", "Mark": "Ford", "Horsepower": 140, "Year": "1970-01-01", "Origin": "USA"},
+           {"Model": "DS-21 Pallas", "Mark": "Citroen", "Horsepower": 115, "Year": "1970-01-01", "Origin": "Europe"},
+           {"Model": "Mustang gl", "Mark": "Ford", "Horsepower": 86, "Year": "1982-01-01", "Origin": "USA"},
+           {"Model": "VW", "Mark": "Pickup", "Horsepower": 52, "Year": "1982-01-01", "Origin": "Europe"}]
