--- week5/lesson1_Flask/homework/DmytroMelnyk/TASK1/app_fabrik/app.py
@@ -1,6 +1,0 @@
-from app_fabrik import run_app
-
-app = run_app()
-
-if __name__ == "__main __":
-    app.run(debug=True)
