--- week5/lesson1_Flask/homework/DmytroMelnyk/TASK1/templates/about.html
@@ -1,4 +1,0 @@
-{% extends "layout.html" %}
-{% block content %}
-     <h1>REST API server</h1>
-{% endblock content %}
