--- week5/lesson1_Flask/homework/DmytroMelnyk/TASK1/templates/home.html
@@ -1,6 +1,0 @@
-{% extends "layout.html" %}
-{% block content %}
-    <h1>Home Page</h1>
-    <a href="/cars">List of Cars Page</a>
-
-{% endblock content %}
